# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards API
  description: REST API for bank card management, user administration and transfers between cards.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Registration and login
  - name: Cards
    description: Card management
  - name: Transfers
    description: Transfers between own cards
  - name: Users
    description: User administration (ADMIN only)
  - name: Block Requests
    description: Card block requests (USER creates, ADMIN processes)

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Creates a user with the USER role.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/cards:
    post:
      tags: [Cards]
      summary: Create a card
      description: ADMIN only. Card number is encrypted in the database; API returns masked number.
      operationId: createCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Cards]
      summary: List all cards
      description: ADMIN only. Supports filtering by status and search by last four digits.
      operationId: getAllCards
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/statusFilter'
        - $ref: '#/components/parameters/searchLastFour'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated list of cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/cards/my:
    get:
      tags: [Cards]
      summary: List current user's cards
      description: Returns cards owned by the authenticated user. Supports filtering by status and search by last four digits.
      operationId: getMyCards
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/statusFilter'
        - $ref: '#/components/parameters/searchLastFour'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated list of cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/cards/{id}:
    get:
      tags: [Cards]
      summary: Get card by id
      description: USER can access own cards only; ADMIN can access any card.
      operationId: getCardById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Cards]
      summary: Delete a card
      description: ADMIN only.
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '204':
          description: Card deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/cards/{id}/block:
    patch:
      tags: [Cards]
      summary: Block a card
      description: ADMIN only.
      operationId: blockCard
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '200':
          description: Card blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/cards/{id}/activate:
    patch:
      tags: [Cards]
      summary: Activate a card
      description: ADMIN only. Expired cards remain EXPIRED.
      operationId: activateCard
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '200':
          description: Card activated or marked expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/cards/{id}/block-request:
    post:
      tags: [Cards]
      summary: Request card block
      description: USER only. Creates a pending block request; card remains active until ADMIN approves.
      operationId: requestBlock
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cardId'
      responses:
        '201':
          description: Block request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/block-requests:
    get:
      tags: [Block Requests]
      summary: List block requests
      description: ADMIN only. Supports filtering by status.
      operationId: getAllBlockRequests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/BlockRequestStatus'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated list of block requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBlockRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/block-requests/{id}/approve:
    patch:
      tags: [Block Requests]
      summary: Approve block request
      description: ADMIN only. Blocks the card and marks the request as approved.
      operationId: approveBlockRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/blockRequestId'
      responses:
        '200':
          description: Block request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/block-requests/{id}/reject:
    patch:
      tags: [Block Requests]
      summary: Reject block request
      description: ADMIN only. Rejects the request; card status is unchanged.
      operationId: rejectBlockRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/blockRequestId'
      responses:
        '200':
          description: Block request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/transfers:
    post:
      tags: [Transfers]
      summary: Transfer between own cards
      operationId: transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/transfers/my:
    get:
      tags: [Transfers]
      summary: List current user's transfers
      operationId: getMyTransfers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated list of transfers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTransferResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/users:
    get:
      tags: [Users]
      summary: List all users
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create a user
      description: ADMIN only. Allows assigning ADMIN or USER role.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/users/{id}/status:
    patch:
      tags: [Users]
      summary: Enable or disable a user
      operationId: updateUserStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: User status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/auth/login or /api/auth/register

  parameters:
    cardId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    userId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    blockRequestId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    statusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/CardStatus'
    searchLastFour:
      name: search
      in: query
      required: false
      description: Last four digits of the card number (exactly 4 digits).
      schema:
        type: string
        pattern: '^\d{4}$'
        example: '1111'
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    sort:
      name: sort
      in: query
      required: false
      description: Spring Data sort parameter, e.g. `createdAt,desc`.
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]

    BlockRequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]

    BlockRequestResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardId:
          type: integer
          format: int64
        maskedNumber:
          type: string
          example: '**** **** **** 1111'
        status:
          $ref: '#/components/schemas/BlockRequestStatus'
        requestedByUsername:
          type: string
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
        processedByUsername:
          type: string
          nullable: true

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 255

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 255

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
        username:
          type: string

    CreateCardRequest:
      type: object
      required: [ownerId, cardNumber, expiryDate]
      properties:
        ownerId:
          type: integer
          format: int64
        cardNumber:
          type: string
          description: Full card number; validated with Luhn algorithm.
        expiryDate:
          type: string
          format: date
        balance:
          type: number
          format: decimal
          minimum: 0
          default: 0

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
          example: '**** **** **** 1111'
        ownerUsername:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time

    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: decimal
          minimum: 0.01

    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        fromCardMaskedNumber:
          type: string
        toCardMaskedNumber:
          type: string
        amount:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 255
        role:
          type: string
          enum: [ADMIN, USER]

    UpdateUserStatusRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time

    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PageTransferResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransferResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PageBlockRequestResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BlockRequestResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PageUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string

    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: Validation failed
        fields:
          type: object
          additionalProperties:
            type: string